[BUGFIX] Consider new universal tag attribute behaviour of `readonly`
The dedicated tag attribute `readonly` of the f:form.textarea ViewHelper was removed by commit https://github.com/TYPO3/typo3/commit/21d3f18b0ee8c7e2915f2ca3d054704ce06a4ec4 in favour of the new universal tag attributes, which allow to define any attributes that are to be output directly in HTML.

Due to the boolean property of the `readonly` attribute, this results in a difference in the output behaviour by Fluid. Previously, an empty value for `readonly=""` in ViewHelper ensured that the `readonly` attribute was not output in HTML, while `readonly="1"` in HTML ensured that the `readonly` attribute was used. Since the changeover in TYPO3 v13, both spellings (with or without a value) result in the `readonly` attribute being output in HTML.

This is now corrected by no longer considering the originally intended conditions for the usage of the `readonly` attribute inline but in the surrounding context.

Resolves: rrrapha/translate_locallang#81